requirements.txt
alembic==1.11.1
bidict==0.22.1
blinker==1.6.2
click==8.1.3
Flask==2.3.2
Flask-Cors==4.0.0
flask-marshmallow==0.15.0
Flask-Migrate==4.0.4
Flask-SocketIO==5.3.4
Flask-SQLAlchemy==3.0.5
h11==0.14.0
importlib-metadata==6.7.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.3
marshmallow==3.19.0
marshmallow-sqlalchemy==0.29.0
packaging==23.1
psycopg2-binary==2.9.6
PyJWT==2.7.0
python-dotenv==1.0.0
python-engineio==4.5.0
python-socketio==5.8.0
simple-websocket==0.10.1
SQLAlchemy==2.0.17
typing_extensions==4.7.1
Werkzeug==2.3.6
wsproto==1.2.0
zipp==3.15.0
